moved sortWallets to calculateMessageSignature
